Split routing views into ember-routing-views
