+++ lib/packages.js
@@ -13,6 +13,7 @@ module.exports = {
   'ember-routing':              {trees: null,  vendorRequirements: ['router', 'route-recognizer'],
                                                requirements: ['ember-runtime', 'ember-views']},
   'ember-routing-handlebars':   {trees: null,  requirements: ['ember-routing', 'ember-handlebars']},
+  'ember-routing-views':        {trees: null,  requirements: ['ember-routing']},
   'ember-application':          {trees: null,  vendorRequirements: ['dag-map'], requirements: ['ember-routing']},
   'ember':                      {trees: null,  requirements: ['ember-application']}
 };
