+++ packages/ember-routing-views/lib/main.js
@@ -1,0 +1,17 @@
+/**
+Ember Routing Views
+
+@module ember
+@submodule ember-routing-views
+@requires ember-routing
+*/
+
+import Ember from "ember-metal/core";
+
+import { LinkView } from "ember-routing-views/views/link";
+import { OutletView } from "ember-routing-views/views/outlet";
+
+Ember.LinkView = LinkView;
+Ember.OutletView = OutletView;
+
+export default Ember;
