+++ packages/ember-routing-views/lib/views/outlet.js
@@ -1,0 +1,9 @@
+/**
+@module ember
+@submodule ember-routing-views
+*/
+
+import ContainerView from "ember-views/views/container_view";
+import { _Metamorph } from "ember-views/views/metamorph_view";
+
+export var OutletView = ContainerView.extend(_Metamorph);
