+++ packages/ember-routing/lib/system/query_params.js
@@ -1,0 +1,6 @@
+import EmberObject from 'ember-runtime/system/object';
+
+export default EmberObject.extend({
+  isQueryParams: true,
+  values: null
+});
