+++ yuidoc.json
@@ -12,6 +12,7 @@
       "packages/ember-handlebars/lib",
       "packages/ember-routing/lib",
       "packages/ember-routing-handlebars/lib",
+      "packages/ember-routing-views/lib",
       "packages/ember-application/lib",
       "packages/ember-testing/lib",
       "packages/ember-handlebars-compiler/lib",
